Moved AutoCompleteinput and fixed some behaviour errors.
