+++ admin/src/customRoutes.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Route } from 'react-router-dom';
 
-import OIDCCallback from './OIDCCallback';
+import OIDCCallback from './auth/OIDCCallback';
 
 export default [
     <Route exact path="/oidc/callback" component={OIDCCallback} noLayout />,
